Best_Buy/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1
django-cors-headers==3.5.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
icalendar==4.0.6
idna==2.10
numpy==1.19.1
pandas==1.1.1
psycopg2==2.8.5
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
schedule==0.6.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
